# api/rag-api-openapi.json
{"openapi":"3.1.0","info":{"title":"RAG Agent API","description":"\n    ## 生产级 RAG + Agent API 服务\n    \n    ### 核心功能：\n    - **文档管理**：上传、向量化、批量导入文档\n    - **智能检索**：基于pgvector的语义搜索\n    - **认证系统**：支持 API Key 和 JWT 双认证\n    - **配额控制**：免费/付费/管理员三级权限\n    \n    ### 技术栈：\n    FastAPI · PostgreSQL+pgvector · Redis · Docker · Prometheus+Grafana\n    ","contact":{"name":"hwd","url":"https://github.com/%E4%BD%A0%E7%9A%84%E7%94%A8%E6%88%B7%E5%90%8D"},"license":{"name":"MIT"},"version":"1.0.0"},"paths":{"/metrics":{"get":{"summary":"Metrics","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["公开"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/create_user":{"post":{"tags":["用户管理"],"summary":"创建新用户并生成API Key","description":"**这个系统内还没有加入初始管理员，此文档仅作为标注记录，参考使用，生成的API Key是30天有效期。**\n    **管理员专用接口**：输入用户名和有效期，生成一个全新的 API Key。\n    \n    **安全注意**：\n    - API Key **仅在此次响应中返回一次**，系统不会存储明文Key。\n    - 请立即复制并妥善保存，关闭页面后将无法找回。\n    - 数据库仅存储 Key 的 SHA256 哈希值，即使数据库泄露也无法还原明文。\n    \n    **认证要求**：当前版本暂未强制校验管理员身份，生产环境中应限制仅管理员可调用。","operationId":"create_user_admin_create_user_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"用户创建成功，返回API Key明文","content":{"application/json":{"schema":{}}}},"422":{"description":"请求参数格式错误，如用户名包含非法字符"},"500":{"description":"服务器内部错误，如数据库写入失败"}}}},"/users/{user_id}":{"get":{"tags":["用户管理"],"summary":"获取用户信息","description":"根据用户ID查询用户基本资料。可选择是否返回详细信息。","operationId":"get_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"description":"用户的唯一数字ID","title":"User Id"},"description":"用户的唯一数字ID","example":123},{"name":"include_detail","in":"query","required":false,"schema":{"type":"boolean","description":"是否返回详细信息（如邮箱、注册时间等），默认仅返回基础信息","default":false,"title":"Include Detail"},"description":"是否返回详细信息（如邮箱、注册时间等），默认仅返回基础信息","example":true}],"responses":{"200":{"description":"查询成功","content":{"application/json":{"schema":{}}}},"422":{"description":"参数格式错误，如 user_id 不是整数"}}}},"/rag/ask":{"post":{"tags":["检索"],"summary":"原有同步接口（需API Key）","operationId":"ask_question_rag_ask_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/async_ask":{"post":{"tags":["检索"],"summary":"异步检索（模拟，无需API Key）","operationId":"async_ask_question_rag_async_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/parallel_ask":{"post":{"tags":["检索"],"summary":"并行检索（模拟）","operationId":"parallel_ask_question_rag_parallel_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/pg_search":{"post":{"tags":["检索"],"summary":"向量语义检索","description":"使用阿里百炼 text-embedding-v2 将问题转为向量，在 pgvector 中检索最相关的文档块，返回按相似度降序排列的结果。","operationId":"pg_search_rag_pg_search_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}}},"responses":{"200":{"description":"检索成功","content":{"application/json":{"schema":{}}}},"401":{"description":"缺少或无效的 API Key"},"403":{"description":"API Key 已过期"},"429":{"description":"请求过于频繁或配额已用完"},"500":{"description":"服务器内部错误"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/insert":{"post":{"tags":["文档管理"],"summary":"插入单条文档","description":"接收一段文本，自动调用阿里百炼 text-embedding-v2 模型将其转换为 1536 维向量，\n    然后存入 PostgreSQL(pgvector) 数据库。\n    \n    文档入库后即可被 `/rag/pg_search` 接口检索。\n    \n    **认证要求**：需要在请求头中携带 `X-API-Key` 或 `Authorization: Bearer <JWT>`。","operationId":"insert_single_doc_rag_insert_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentInsert"}}}},"responses":{"200":{"description":"文档插入成功","content":{"application/json":{"schema":{}}}},"401":{"description":"缺少或无效的 API Key / JWT"},"403":{"description":"API Key 已过期或无权访问"},"429":{"description":"请求过于频繁或今日配额已用完"},"500":{"description":"服务器内部错误（如 Embedding API 调用失败）"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/insert_batch":{"post":{"tags":["文档管理"],"summary":"批量插入文档","description":"批量插入文档，自动调用阿里百炼 text-embedding-v2 模型将其转换为 1536 维向量，\n    然后存入 PostgreSQL(pgvector) 数据库。\n    \n    文档入库后即可被 `/rag/pg_search` 接口检索。\n    \n    **认证要求**：需要在请求头中携带 `X-API-Key` 或 `Authorization: Bearer <JWT>`。","operationId":"insert_batch_docs_rag_insert_batch_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDocumentInsert"}}}},"responses":{"200":{"description":"文档插入成功","content":{"application/json":{"schema":{}}}},"401":{"description":"缺少或无效的 API Key / JWT"},"403":{"description":"API Key 已过期或无权访问"},"429":{"description":"请求过于频繁或今日配额已用完"},"500":{"description":"服务器内部错误（如 Embedding API 调用失败）"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/documents/{doc_id}":{"delete":{"tags":["文档管理"],"summary":"删除指定文档","description":"根据文档的唯一数字ID，从 PostgreSQL(pgvector) 数据库中永久删除该文档记录。\n    \n    删除后该文档将不再出现在 `/rag/pg_search` 的检索结果中。\n    \n    **认证要求**：需要在请求头中携带 `X-API-Key` 或 `Authorization: Bearer <JWT>`。\n    **权限要求**：仅管理员用户可执行此操作（当前版本暂未强制校验，后续会完善）。","operationId":"delete_document_rag_documents__doc_id__delete","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","minimum":1,"description":"要删除的文档唯一数字ID","title":"Doc Id"},"description":"要删除的文档唯一数字ID","example":1},{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"文档删除成功","content":{"application/json":{"schema":{}}}},"401":{"description":"缺少或无效的 API Key / JWT"},"403":{"description":"API Key 已过期或无权访问"},"404":{"description":"指定 ID 的文档不存在"},"422":{"description":"参数格式错误，如 doc_id 不是整数"},"429":{"description":"请求过于频繁或今日配额已用完"},"500":{"description":"服务器内部错误"}}}},"/rag/benchmark-embedding":{"post":{"tags":["调试"],"summary":"对比测试：缓存命中和无缓存命中 时间差距 ，正常生产级数据库大概是5倍，看数据库大小","operationId":"benchmark_embedding_rag_benchmark_embedding_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/cache_stats":{"get":{"tags":["调试"],"summary":"调试接口，用来查看当前缓存中有多少个 Embedding 键","operationId":"cache_stats_debug_cache_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/tool/benchmark":{"get":{"tags":["调试"],"summary":"调试接口，对比缓存前后的工具调用耗时s","description":"对比缓存前后的工具调用耗时","operationId":"benchmark_tool_tool_benchmark_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/debug/rate_limit/{user_name}":{"get":{"tags":["调试"],"summary":"调试接口，查询某用户的剩余令牌数","description":"查询某用户的剩余令牌数","operationId":"check_rate_limit_debug_rate_limit__user_name__get","parameters":[{"name":"user_name","in":"path","required":true,"schema":{"type":"string","title":"User Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["认证"],"summary":"获取短期令牌和长期令牌","description":"用户登录，返回 access_token短期令牌 和 refresh_token长期令牌","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"新创建的短期令牌时效15分钟，长期令牌30天","content":{"application/json":{"schema":{}}}},"401":{"description":"用户名或密码错误"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["认证"],"summary":"换取新的短期令牌","description":"输入refresh_token，用 refresh_token 换取新的 access_token","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"新创建的短期令牌时效15分钟","content":{"application/json":{"schema":{}}}},"401":{"description":"无效或过期的刷新令牌"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/jwt_ask":{"post":{"tags":["检索"],"summary":"JWT认证的问答接口","description":"输入JWT_SECRET_KEY，获得回答结果","operationId":"jwt_ask_question_rag_jwt_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionRequest"}}},"required":true},"responses":{"200":{"description":"返回搜索结果：创建JWT_SECRET_KEY，用户名，下的数据结果","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/debug/quota/{user_name}":{"get":{"tags":["调试"],"summary":"权限分流的测试接口","description":"权限分流的测试接口，限流（按角色不同限额）","operationId":"check_quota_debug_quota__user_name__get","parameters":[{"name":"user_name","in":"path","required":true,"schema":{"type":"string","title":"User Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/debug/count":{"get":{"tags":["调试"],"summary":"调试：查看数据库中文档数量","operationId":"debug_count_debug_count_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BatchDocumentInsert":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/DocumentInsert"},"type":"array","title":"Documents","description":"要批量插入的文档列表","examples":[[{"content":"Python是一种广泛使用的编程语言，常用于数据分析和AI开发。","source":"tech_docs"},{"content":"今天中午吃什么呢？这是一个让人纠结的哲学问题。","source":"life_docs"}]]}},"type":"object","required":["documents"],"title":"BatchDocumentInsert","examples":[{"documents":[{"content":"Python是一种广泛使用的编程语言，常用于数据分析和AI开发。","source":"tech_docs"},{"content":"今天中午吃什么呢？这是一个让人纠结的哲学问题。","source":"life_docs"},{"content":"FastAPI是一个高性能的Python Web框架，支持异步处理。","source":"tech_docs"}]}]},"DocumentInsert":{"properties":{"content":{"type":"string","title":"Content","description":"要入库的文档文本内容","example":"Python是一种广泛使用的编程语言。"},"source":{"type":"string","title":"Source","description":"文档来源标识，如 tech_docs, life_docs","default":"manual","example":"tech_docs"}},"type":"object","required":["content"],"title":"DocumentInsert"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"user_name":{"type":"string","title":"User Name","description":"用户名","example":"admin"},"password":{"type":"string","title":"Password","description":"密码","example":"admin123"}},"type":"object","required":["user_name","password"],"title":"LoginRequest"},"QuestionRequest":{"properties":{"question":{"type":"string","title":"Question","description":"用户输入的自然语言问题","example":"FastAPI是什么？"},"top_k":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Top K","description":"返回的最相似文档数量","default":3,"example":5}},"type":"object","required":["question"],"title":"QuestionRequest","example":{"question":"FastAPI是什么？","top_k":5}},"RefreshRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token","description":"用于获取新访问令牌的刷新令牌","examples":["eyJhbGciOiJIUzI1NiIs..."]}},"type":"object","required":["refresh_token"],"title":"RefreshRequest","examples":[{"refresh_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}]},"UserCreate":{"properties":{"user_name":{"type":"string","title":"User Name","description":"用户名，用于标识调用者","example":"test_user"},"expire_days":{"type":"integer","maximum":365.0,"minimum":1.0,"title":"Expire Days","description":"API Key 的有效天数","default":30,"example":90}},"type":"object","required":["user_name"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}